// src/components/molecules/maps/maps.d.ts
export interface MapsProps {
  position?: [number, number];
  zoom?: number;
  popUpText?: string;
  className?: string;
  setChoosenLocation?: (data: ILonLat) => void;
  choosenLocation?: ILonLat;
}

export interface ILonLat {
  lon: string;
  lat: string;
  display_name: string;
  name?: string;
}